[#438] MOVE: 패키지 시그니처 변경
