+++ backend/src/main/java/com/team5/secondhand/application/chatroom/controller/ChatroomController.java
@@ -6,12 +6,11 @@ import com.team5.secondhand.application.chatroom.dto.response.ChatroomList;
 import com.team5.secondhand.application.chatroom.exception.BuyerException;
 import com.team5.secondhand.application.chatroom.exception.ExistChatRoomException;
 import com.team5.secondhand.application.chatroom.exception.NotChatroomMemberException;
-import com.team5.secondhand.application.chatroom.service.ChatroomFacade;
 import com.team5.secondhand.application.item.exception.ExistItemException;
 import com.team5.secondhand.application.member.dto.response.MemberDetails;
 import com.team5.secondhand.application.member.exception.ExistMemberIdException;
 import com.team5.secondhand.global.auth.Login;
-import com.team5.secondhand.global.dto.GenericResponse;
+import com.team5.secondhand.global.model.GenericResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
