+++ backend/src/main/java/com/team5/secondhand/application/item/controller/v2/ItemControllerV2.java
@@ -4,7 +4,7 @@ import com.team5.secondhand.application.item.controller.v2.dto.ItemsRequest;
 import com.team5.secondhand.application.item.controller.v2.dto.ItemsResponse;
 import com.team5.secondhand.application.member.dto.response.MemberDetails;
 import com.team5.secondhand.application.region.exception.NotValidRegionException;
-import com.team5.secondhand.global.dto.GenericResponse;
+import com.team5.secondhand.global.model.GenericResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
